Cart Total
This required a bit of figuring out. Calling this.setState() then another function to calculate the cart total, and then this.setState() can cause errors due to delays in the state update.

To fix this UpdateBasketTotal() needs to be given data, and returns the total, so that only one this.setState() is required.